[Feat] header 계정 관리 이동 link 추가
