+++ src/components/common/Header.tsx
@@ -13,7 +13,7 @@ import { getMember } from "@/api/member";
 import { Modal } from "./ModalPopup";
 import InputModal from "../ModalContents/InputModal";
 import { createInvite } from "@/api/dashboard";
-import Logout from "./Dropdown/LogoutDropdown";
+import ProfileSelect from "./Dropdown/Profile.D.D";
 
 interface Props {
   createdByMe?: boolean;
@@ -25,7 +25,6 @@ export default function Header({ createdByMe }: Props) {
   const router = useRouter();
   const [members, setMembers] = useState<Member[]>([]);
   const [inviteValue, setInviteValue] = useState("");
-  const display = router.pathname === "/mydashbord" ? "none" : "block";
   const dashboardTitle =
     router.pathname === "/mypage"
       ? "계정관리"
@@ -52,7 +51,7 @@ export default function Header({ createdByMe }: Props) {
 
   //
   return (
-    <div className="flex bg-white flex-row justify-between w-full h-[70px] py-[15px] px-[8px] tablet:pl-10 tablet:pl-[74px] tablet:pr-[32px] laptop:pr-[80px] border-b-[1px] items-center border-gray-300 ">
+    <div className="flex bg-white flex-row justify-between w-full h-[70px] py-[15px] px-[8px] tablet:pl-[74px] tablet:pr-[32px] laptop:pr-[80px] border-b-[1px] items-center border-gray-300 ">
       <div className="flex flex-row items-center gap-2">
         <p className="hidden laptop:block text-black-200 text-xl-bold ">
           {dashboardTitle}
@@ -65,20 +64,22 @@ export default function Header({ createdByMe }: Props) {
       </div>
       <div className="flex flex-row gap-[16px] tablet:gap-[36px] laptop:gap-[40px]">
         <div className="flex flex-row gap-2 tablet:gap-4">
-          <Link href="/mypage">
-            <button className="justify-center h-[40px] flex flex-row items-center py-[7px] px-2 tablet:px-4 gap-2 rounded-lg border-[1px] border-gray-300">
-              <Image
-                className="hidden tablet:block"
-                src={Edit}
-                width={18}
-                height={18}
-                alt="톱니바퀴"
-              />
-              <div className="text-gray-500 text-md-medium tablet:text-lg-medium">
-                관리
-              </div>
-            </button>
-          </Link>
+          {router.pathname !== "/mydashboard" && (
+            <Link href={`${dashboardId}/edit`}>
+              <button className="justify-center h-[40px] flex flex-row items-center py-[7px] px-2 tablet:px-4 gap-2 rounded-lg border-[1px] border-gray-300">
+                <Image
+                  className="hidden tablet:block"
+                  src={Edit}
+                  width={18}
+                  height={18}
+                  alt="톱니바퀴"
+                />
+                <div className="text-gray-500 text-md-medium tablet:text-lg-medium">
+                  관리
+                </div>
+              </button>
+            </Link>
+          )}
           {router.pathname !== "/mydashboard" && (
             <div className="h-[40px]">
               <Modal
@@ -111,20 +112,22 @@ export default function Header({ createdByMe }: Props) {
           )}
         </div>
         <div className="flex w-fit flex-row gap-[4px] tablet:gap-[36px] laptop:gap-[40px]">
-        <div
-          style={{ display: members && members.length > 0 ? "block" : "none" }}
-        >
-          {members && <Badges memberList={members} />}
-        </div>
+          <div
+            style={{
+              display: members && members.length > 0 ? "block" : "none",
+            }}
+          >
+            {members && <Badges memberList={members} />}
+          </div>
           <div className="h-[38px] w-[1px] bg-gray-300" />
           {store && (
-            <Logout>
+            <ProfileSelect>
               <Profile
                 nickname={store.userNickname}
                 profileImageUrl={store.profileImageUrl}
                 type="profile"
               />
-            </Logout>
+            </ProfileSelect>
           )}
         </div>
       </div>
